% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_tables.R
\name{summarize_shap_rslt}
\alias{summarize_shap_rslt}
\title{Summarize results from \code{\link[=run_survey_shap]{run_survey_shap()}}}
\usage{
summarize_shap_rslt(shap_rslt, top_n = 5L, quiet = FALSE)
}
\arguments{
\item{shap_rslt}{A result list from \code{\link[=run_survey_shap]{run_survey_shap()}}.}

\item{top_n}{Number of top and bottom rows to print for direction tables.}

\item{quiet}{Logical; if TRUE, do not print.}
}
\value{
An object of class \code{surveySHAP_summary}.
}
\description{
Summarize results from \code{\link[=run_survey_shap]{run_survey_shap()}}
}
